# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'ФудГрам'
  version: ''
paths:
...
  /api/recipes/{id}/:
    get:
      operationId: Получение рецепта
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: ''
      tags:
        - Рецепты
    patch:
      operationId: Обновление рецепта
      security:
        - Token: [ ]
      description: 'Доступно только автору данного рецепта'
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта."
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreateUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeList'
          description: 'Рецепт успешно обновлен'
        '400':
          $ref: '#/components/responses/NestedValidationError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
    delete:
      operationId: Удаление рецепта

      description: 'Доступно только автору данного рецепта'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого рецепта"
          schema:
            type: string
      responses:
        '204':
          description: 'Рецепт успешно удален'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionDenied'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Рецепты
...
components:
  schemas:
    User:
      description:  'Пользователь (В рецепте - автор рецепта)'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Вася"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Пупкин"
        is_subscribed:
          type: boolean
          readOnly: true
          description: "Подписан ли текущий пользователь на этого"
          example: false
      required:
        - username
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Название'
          example: 'Завтрак'
        color:
          type: string
          nullable: true
          maxLength: 7
          description: 'Цвет в HEX'
          example: '#E26C2D'
        slug:
          type: string
          nullable: true
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
          description: 'Уникальный слаг'
          example: 'breakfast'
    RecipeList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        tags:
          description: 'Список тегов'
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        author:
          $ref: '#/components/schemas/User'
        ingredients:
          description: 'Список ингредиентов'
          type: array
          items:
            $ref: '#/components/schemas/IngredientInRecipe'
        is_favorited:
          type: boolean
          description: 'Находится ли в избранном'
        is_in_shopping_cart:
          type: boolean
          description: 'Находится ли в корзине'
        name:
          type: string
          maxLength: 200
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://foodgram.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - tags
        - author
        - is_favorited
        - is_in_shopping_cart
        - name
        - image
        - text
        - cooking_time
...
    IngredientInRecipe:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
          description: 'Название'
          example: 'Картофель отварной'
        measurement_unit:
          type: string
          maxLength: 200
          description: 'Единицы измерения'
          example: 'г'
        amount:
          type: integer
          description: 'Количество'
          minimum: 1

      required:
        - name
        - measurement_unit
...
    RecipeCreateUpdate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        ingredients:
          description: Список ингредиентов
          type: array
          items:
            example:
              id: 1123
              amount: 10
            type: object
            properties:
              id:
                description: 'Уникальный id'
                type: integer
              amount:
                description: 'Количество в рецепте'
                type: integer
            required:
              - id
              - amount
        tags:
          description: 'Список id тегов'
          type: array
          example: [1, 2]
          items:
            type: integer
        image:
          description: 'Картинка, закодированная в Base64'
          example: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABAgMAAABieywaAAAACVBMVEUAAAD///9fX1/S0ecCAAAACXBIWXMAAA7EAAAOxAGVKw4bAAAACklEQVQImWNoAAAAggCByxOyYQAAAABJRU5ErkJggg=='
          type: string
          format: binary
        name:
          description: 'Название'
          type: string
          maxLength: 200
        text:
          description: 'Описание'
          type: string
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
      required:
        - ingredients
        - tags
        - image
        - name
        - text
        - cooking_time
    ValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        field_name:
          description: 'Название поля, в котором произошли ошибки. Таких полей может быть несколько'
          example: [ 'Обязательное поле.' ]
          type: array
          items:
            type: string
    NestedValidationError:
      description: Стандартные ошибки валидации DRF
      type: object
      properties:
        ingredients:
          description: 'Ошибки в ингредиентах. В приведенном примере в первом и третьем ингредиенте не было ошибок (amount >= 1), а во втором были.'
          example: [{},{"amount":["Убедитесь, что это значение больше либо равно 1."]}, {}]
          type: array
          items:
            type: object
            properties:
              amount:
                type: array
                items:
                  type: string
...
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
    PermissionDenied:
      description: Недостаточно прав
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "У вас недостаточно прав для выполнения данного действия."
          type: string
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
...
    NestedValidationError:
      description: 'Ошибки валидации в стандартном формате DRF, в том числе с внутренними элементами.'
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/NestedValidationError'
              - $ref: '#/components/schemas/ValidationError'
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
    PermissionDenied:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PermissionDenied'
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token